fixed Controller.java and Search.java
Changed searchByTitle and searchByID methods in Search.java to return Movies object instead of string, also fixed misnamed variable.

In controller.java I commented out the title string in the handleButton method as it would cause compile error as it was trying to assign a movies object to a string.

Program should be able to launch successfully now.